requirements.txt
alembic==1.11.1
aniso8601==8.0.0
APScheduler==3.10.1
astroid==2.4.2
async-timeout==4.0.2
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.6.2
cachelib==0.8.0
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
click==8.1.3
coverage==5.3
cryptography==36.0.0
cx-Oracle==8.3.0
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Caching==1.11.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
Flask-SSE==1.0.0
greenlet==2.0.2
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.3
ldap3==2.9.1
Mako==1.1.3
MarkupSafe==2.1.3
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mccabe==0.6.1
packaging==21.3
paramiko==2.8.0
pdfkit==1.0.0
prettytable==2.4.0
psycopg2-binary
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==2.2.0
PyJWT==2.7.0
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==3.0.9
python-barcode==0.14.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
redis==4.5.5
requests==2.25.0
rq==1.15.1
rq-scheduler==0.13.1
six==1.15.0
SQLAlchemy==2.0.15
sshtunnel==0.4.0
toml==0.10.2
typing_extensions==4.6.3
tzdata==2022.7
tzlocal==4.3
unittest-xml-reporting==3.0.4
urllib3==1.26.2
wcwidth==0.2.5
webargs==8.2.0
Werkzeug==2.3.6
wrapt==1.12.1
XlsxWriter==3.1.0
Office365-REST-Python-Client==2.4.3
